--- packages/anvil-ft-app-context/browser.js
@@ -1,1 +1,0 @@
-export * from './dist/browser/client/index'
