--- packages/anvil-ft-app-context/server.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/node/server/index')
