+++ packages/anvil-ft-app-context/src/browser.ts
@@ -1,0 +1,7 @@
+import { AppContextClient } from './clients/AppContextClient'
+import { loadEmbeddedAppContextData } from './embedding/loadEmbeddedAppContextData'
+
+export function init() {
+  const context = loadEmbeddedAppContextData()
+  return new AppContextClient({ context })
+}
