--- packages/anvil-ft-app-context/src/client/index.ts
@@ -1,2 +1,0 @@
-export { default as init } from './init'
-export { default as AppContextClient } from '../shared/appContext/AppContextClient'
