--- packages/anvil-ft-app-context/src/client/init.ts
@@ -1,7 +1,0 @@
-import AppContext from '../shared/appContext/AppContextClient'
-import loadAppContextData from './loadAppContextData'
-
-export default function init() {
-  const context = loadAppContextData()
-  return new AppContext({ context })
-}
