--- packages/anvil-ft-app-context/src/client/loadAppContextData.ts
@@ -1,7 +1,0 @@
-import { TAppContext } from '../types'
-import loadFromScriptEmbed from './loadFromScriptEmbed'
-import loadFromDataAttributesEmbed from './loadFromDataAttributesEmbed'
-
-export default function loadAppContextData(): TAppContext {
-  return loadFromScriptEmbed() || loadFromDataAttributesEmbed()
-}
