+++ packages/anvil-ft-app-context/src/embedding/loadEmbeddedAppContextData.ts
@@ -1,0 +1,7 @@
+import { TAppContext } from '../types'
+import { loadDataFromScriptEmbed } from './loadDataFromScriptEmbed'
+import { loadDataFromAttributesEmbed } from './loadDataFromAttributesEmbed'
+
+export function loadEmbeddedAppContextData(): TAppContext {
+  return loadDataFromScriptEmbed() || loadDataFromAttributesEmbed()
+}
