+++ packages/anvil-ft-app-context/src/schema/typeOfSchemaProp.ts
@@ -1,0 +1,5 @@
+import schema from './schema.json'
+
+export function typeOfSchemaProp(propName: string) {
+  return schema.properties[propName].type
+}
