--- packages/anvil-ft-app-context/src/shared/appContext/index.ts
@@ -1,5 +1,0 @@
-import * as constants from './constants'
-
-export { constants }
-export { default as AppContextClient } from './AppContextClient'
-export { default as prepareEmbedString } from './prepareEmbedString'
