--- packages/anvil-ft-app-context/src/shared/appContext/prepareEmbedString.ts
@@ -1,7 +1,0 @@
-import { TAppContext } from '../../types'
-import { APP_CONTEXT_ELEMENT_ID } from './constants'
-
-export default function prepareEmbedString(appContext: TAppContext) {
-  const jsonData = JSON.stringify(appContext)
-  return `<script type="application/json" id="${APP_CONTEXT_ELEMENT_ID}">${jsonData}</script>`
-}
